requirements.txt
fastapi
gunicorn
jupyter
pandas
pypyr
python-dotenv
tensorflow
uvicorn
boto3
pydantic
dataclasses
pymongo
python-dotenv